+++ src/core/injections/controller/checks.py
@@ -856,6 +856,15 @@ def check_quotes(payload):
 Recognise the payload.
 """
 def recognise_payload(payload):
+  if "usleep" in payload:
+    if not settings.TAMPER_SCRIPTS['sleep2usleep']:
+      if menu.options.tamper:
+        menu.options.tamper = menu.options.tamper + ",sleep2usleep"
+      else:
+        menu.options.tamper = "sleep2usleep"  
+    from src.core.tamper import sleep2usleep
+    payload = sleep2usleep.transform(payload)
+
   is_decoded = False
   if (len(payload) % 4 == 0) and \
     re.match(settings.BASE64_RECOGNITION_REGEX, payload) and \
@@ -905,6 +914,10 @@ Perform payload modification
 """
 def perform_payload_modification(payload):
   for encode_type in list(set(settings.MULTI_ENCODED_PAYLOAD[::-1])):
+    # sleep to usleep
+    if encode_type == 'sleep2usleep':
+      from src.core.tamper import sleep2usleep
+      payload = sleep2usleep.transform(payload)
     # Add single quotes.
     if encode_type == 'singlequotes':
       from src.core.tamper import singlequotes