+++ src/utils/settings.py
@@ -111,7 +111,7 @@ APPLICATION = "commix"
 DESCRIPTION_FULL = "Automated All-in-One OS Command Injection and Exploitation Tool"
 DESCRIPTION = "The command injection exploiter"
 AUTHOR  = "Anastasios Stasinopoulos"
-VERSION_NUM = "2.6.7"
+VERSION_NUM = "2.6.8"
 STABLE_VERSION = False
 if STABLE_VERSION:
   VERSION = "v" + VERSION_NUM[:3] + "-stable"
@@ -685,7 +685,8 @@ TAMPER_SCRIPTS = {
                   "caret": False,
                   "multiplespaces":False,
                   "backslashes":False,
-                  "nested":False
+                  "nested":False,
+                  "sleep2usleep":False
                  }
 
 # HTTP Errors